colocate screens in route folder, also collapse into page if page had no additional logic
