+++ src/app/play/[date]/EndScreen.tsx
